pyproject.toml
[project]
name = "analyticsbot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "aiogram>=3.24.0",
    "asyncpg>=0.31.0",
    "asyncpg-stubs>=0.31.1",
    "loguru>=0.7.3",
    "openai>=2.16.0",
    "pydantic-settings>=2.12.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
]
